add cursor.cursorAt(fullPathArr)
Add ability to generate subpath cursors using a fullPath array. This makes it easier to reason about a path since you don't need to discern and piece together relative paths from all the parents.

It allows for easier maintenance of structures.

Current implementation throws an error if you request a path that is not equal or subpath of existing which is ideal for top down subcursor flows.